refactor: tidy up
