+++ apps/extension/src/ui/apps/dashboard/routes/AccountAdd/AccountAddLedgerWizard.tsx
@@ -1,7 +1,6 @@
+import { DashboardLayout } from "@ui/apps/dashboard/layout/DashboardLayout"
 import { useSelectAccountAndNavigate } from "@ui/hooks/useSelectAccountAndNavigate"
 import { lazy } from "react"
-
-import { AccountAddLedgerLayout } from "./AccountAddLedgerLayout"
 
 const AccountAddLedgerWizard = lazy(() => import("@ui/domains/Account/AccountAdd/AccountAddLedger"))
 
@@ -9,8 +8,8 @@ export const AccountAddLedgerDashboardWizard = () => {
   const { setAddress } = useSelectAccountAndNavigate("/portfolio")
 
   return (
-    <AccountAddLedgerLayout>
+    <DashboardLayout withBack centered>
       <AccountAddLedgerWizard onSuccess={setAddress} />
-    </AccountAddLedgerLayout>
+    </DashboardLayout>
   )
 }
