--- apps/extension/src/ui/apps/dashboard/routes/AccountAdd/AccountAddLedgerWizard/AccountAddLedgerLayout.tsx
@@ -1,9 +1,0 @@
-import { DashboardLayout } from "@ui/apps/dashboard/layout/DashboardLayout"
-
-export const AccountAddLedgerLayout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <DashboardLayout withBack centered>
-      {children}
-    </DashboardLayout>
-  )
-}
