--- apps/extension/src/ui/apps/dashboard/routes/AccountAdd/AccountAddSecretWizard/AccountAddSecretLayout.tsx
@@ -1,9 +1,0 @@
-import { DashboardLayout } from "@ui/apps/dashboard/layout/DashboardLayout"
-
-export const AccountAddSecretLayout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <DashboardLayout withBack centered>
-      {children}
-    </DashboardLayout>
-  )
-}
