+++ apps/extension/src/ui/apps/onboard/routes/AddAccount/AccountAddLedgerWizard.tsx
@@ -1,15 +1,17 @@
+import { OnboardDialog } from "@ui/apps/onboard/components/OnboardDialog"
 import { useOnboard } from "@ui/apps/onboard/context"
+import { Layout } from "@ui/apps/onboard/layout"
 import { lazy } from "react"
-
-import { AccountAddLedgerLayout } from "./AccountAddLedgerLayout"
 
 const AccountAddLedgerWizard = lazy(() => import("@ui/domains/Account/AccountAdd/AccountAddLedger"))
 
 export const AccountAddLedgerOnboardWizard = () => {
   const { setOnboarded } = useOnboard()
   return (
-    <AccountAddLedgerLayout>
-      <AccountAddLedgerWizard onSuccess={setOnboarded} />
-    </AccountAddLedgerLayout>
+    <Layout withBack>
+      <OnboardDialog className="w-[68rem]">
+        <AccountAddLedgerWizard onSuccess={setOnboarded} />
+      </OnboardDialog>
+    </Layout>
   )
 }
