+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddQr/ConfigureAccount.tsx
