+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddQr/ConfigureVerifierCertificateMnemonic.tsx
