+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddQr/MnemonicForm.tsx
