+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddQr/Scan.tsx
